semana05/dia5/shop/requirements.txt
django==3.2
Pillow
djangorestframework
django-cors-headers